backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.2.3
dj-database-url==0.5.0
django-heroku==0.3.1
django-phone-field==1.8.0
django==3.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
